chore/feat(cms): rename file and add get service details func
